+++ internal/providers/all/all.go
@@ -3,6 +3,7 @@ package all
 import (
 	_ "github.com/openclaw/crabbox/internal/providers/aws"
 	_ "github.com/openclaw/crabbox/internal/providers/azure"
+	_ "github.com/openclaw/crabbox/internal/providers/azuredynamicsessions"
 	_ "github.com/openclaw/crabbox/internal/providers/blacksmith"
 	_ "github.com/openclaw/crabbox/internal/providers/cloudflare"
 	_ "github.com/openclaw/crabbox/internal/providers/daytona"